+++ examples/drivers/gopass.sh
@@ -1,18 +1,18 @@
 #!/usr/bin/env sh
 
 # shellcheck disable=SC2034
-_DRIVER_REGEX='!gopass [A-Za-z0-9\-\_\/]*'
+_BACKEND_REGEX='!gopass [A-Za-z0-9\-\_\/]*'
 
-# shellcheck source=scripts//drivers/_custom.sh
-. "${SCRIPT_DIR}/drivers/_custom.sh"
+# shellcheck source=scripts/backends/_custom.sh
+. "${SCRIPT_DIR}/backends/_custom.sh"
 
 _gopass() {
     # shellcheck disable=SC2086
-    set -- ${SECRET_DRIVER_ARGS} "$@"
+    set -- ${SECRET_BACKEND_ARGS} "$@"
     gopass "$@"
 }
 
-_custom_driver_get_secret() {
+_custom_backend_get_secret() {
     _type=$1
     _SECRET=$2
 
@@ -23,7 +23,7 @@ _custom_driver_get_secret() {
 
     if ! _gopass show -o "${_SECRET}"; then
         echo "Error while get secret from gopass!" >&2
-        echo gopass show -o "${_SECRET}" "${SECRET_DRIVER_ARGS}" >&2
+        echo gopass show -o "${_SECRET}" "${SECRET_BACKEND_ARGS}" >&2
         exit 1
     fi
 }