+++ scripts/commands/dec.sh
@@ -22,13 +22,13 @@ EOF
 decrypt_helper() {
     encrypted_file_path="${1}"
 
-    if ! driver_is_file_encrypted "${encrypted_file_path}"; then
+    if ! backend_is_file_encrypted "${encrypted_file_path}"; then
         return 1
     fi
 
     encrypted_file_dec="$(_file_dec_name "${encrypted_file_path}")"
 
-    if ! driver_decrypt_file "yaml" "${encrypted_file_path}" "${encrypted_file_dec}"; then
+    if ! backend_decrypt_file "yaml" "${encrypted_file_path}" "${encrypted_file_dec}"; then
         rm -rf "${encrypted_file_dec}"
         fatal 'Error while decrypting file: %s' "${file}"
     fi